Add files via upload
We added in commision (fixed fee) and slippage. Also, we changed the plot to scatter (for the take profit and stop loss) in order to avoid the diagonal line (buying and selling turns those two levels upside down). We also changed the period to 2m with a smaller time frame (otherwise there would be too many datapoints).